### 2025-06-10 #8356 @HeloiseS
For when we have an editor, here is my suggestion for reviewers:

* @VasanthRajendran is a senior software developer at amazon who reached out to me to offer to help review the code. He'd be great to bring an API dev expertise to the review.

* @karllark is an experienced observational astronomer who will be able to feedback on the functionalities themselves and the documentation.
