### 2025-11-18 #9462 @editorialbot
[REVIEW]: A Python client for the ATLAS API
**Submitting author:** @heloises (<a href="http://orcid.org/0000-0002-0504-4323">H. F. Stevance</a>)
**Repository:** https://github.com/HeloiseS/atlasapiclient
**Branch with paper.md** (empty if default branch): main
**Version:** v1.2.0
