### 2025-11-24 #9462 @ygrange
I'm planning to do some real reviewing tomorrow, but I thought I start with the first few check marks and I already have a comment/question:
Seems like the last author, who is the corresponding author as well, did not commit anything into the repository. I don't really know if there is a policy about this at JOSS (probably more a question to the editor). Maybe it would be nice to add a footnote about their contribution?

(also I just requested access to the data)
