### 2025-11-26 #9462 @ygrange
Hi @HeloiseS: Thanks, I got access yesterday (and even thoug my plan was to look at this yesterday while spending a few hours on trains, I ended up meeting colleagues and socialising).

I personally really like the putting it in the README idea (I think there is also the option of adding a CONTRIBUTORS file, but to me that feels a bit like hiding it). Does the API itself also have a repo? If so, maybe a link to there would also be a good idea. Having the corresponding author someone who has a fixed position seems indeed to be a fair choice to make, for what my opinion here is worth :). Also to be clear: I was not questioning anybodies intentions here, just making sure that there is no unclarity here.
