### 2026-01-08 #9462 @VasanthRajendran
Apologies for the delay, got caught up with end of year deliverables at work. I've gone through the codebase and documentation now.

Summary: Good to go from my side. The package does what it claims to do and the code quality is solid.

Minor comments (not blocking):
Type hints would be nice to have for better IDE support but not essential. Maybe add a FAQ or troubleshooting section for token/auth issues? That's usually where people get stuck with API clients.

I've ticked off all items in my checklist. Thanks for the opportunity to review. Interesting to see how the astronomy community handles survey data pipelines.
