### 2026-01-18 #9462 @Romain-Thomas-Shef
Dear @VasanthRajendran

Thank you for your review. Based on your last message I think it would be good to create an issue in the documentation regarding the comment on "troubleshooting section for token/auth issues". Can you create an issue on the repo?
