### 2026-01-18 HeloiseS/atlasapiclient#37 @Romain-Thomas-Shef
Fix reference for JOSS paper
Dear @HeloiseS,

As noted during the review (https://github.com/openjournals/joss-reviews/issues/9462) by @ygrange, one reference should probably be corrected.

> ```
> Heloise, & genghisken. (2025)
> ```

It points to this[ zenodo entry](https://zenodo.org/records/14983116) so I would write down the full name of the authors as displayed in zenodo and provided by the platform bibtex option:

@software{heloise_2025_14983116,
  author       = {Heloise and
                  Smith, Ken W},
  title        = {HeloiseS/atlasvras: VRA 1.1},
  month        = mar,
  year         = 2025,
  publisher    = {Zenodo},
  version      = {v1.1},
  doi          = {10.5281/zenodo.14983116},
  url          = {https://doi.org/10.5281/zenodo.14983116},
  swhid        = {swh:1:dir:d58f7330359b6c76e796cc182ac71d0d1524d84f
                   ;origin=https://doi.org/10.5281/zenodo.14363396;vi
                   sit=swh:1:snp:46746f947d79fd3b95b63ea09ee85fd2e8b3
                   a397;anchor=swh:1:rel:5af41e199f18b3e1baafe237eb35
                   1b4dc7d05a62;path=HeloiseS-atlasvras-f64733f
                  },
}
