### 2026-02-09 HeloiseS/atlasapiclient#37 @HeloiseS
thanks - replaced the outdates citaation
