+++ _sources/users.rst
@@ -1,5 +1,11 @@
+The Docs
+=====================
+
 Set-up
-============
+--------
+
+Install
+~~~~~~~~~
 
 First you need to install the package, you can either use pip
 
@@ -15,7 +21,7 @@ or clone the repository from github:
 
 
 Config file
-~~~~~~~~~
+~~~~~~~~~~~~~
 The client requires a config ile that contains the base url of the ATLAS transient web servers and your token for the ATLAS API.
 In the directory `atlasapiclient/config_files` you will find the `api_config_template.yaml` file.
 
@@ -26,15 +32,21 @@ In the directory `atlasapiclient/config_files` you will find the `api_config_tem
    cd atlasapiclient/config_files
    cp api_config_template.yaml api_config_MINE.yaml
 
+
 .. warning::
-   The name of that file matters immensly. **Unless you know what you're doing DON'T CHANGE IT**
+    As you'll see below we don't explicitely parse the config file because we assume you have named yours  ``api_config_MINE.yaml``.
+    The path to that file is encoded in the ``API_CONFIG_FILE`` variable in the ``atlasapiclient/utils.py`` file and parsed by default to the classes.
+    If you feel fancy and want to name your config file differently you have to keep track of its location and parse it with the argument `api_config_file`.
+
 
 * 2) Update your token (if you don't have a token see below)
 * 3) Update the url to "https://psweb.mp.qub.ac.uk/sne/atlas4/api/"
 
 
+
+
 How do I get a token?
-~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~
 Once you have access to the web server you can get a token for API access by using the 
 `APIClient.refresh_token()` method. This will generate a token for you and save 
 it in the config file defined within the `API_CONFIG_FILE` variable in the `atlasapiclient/utils.py` file.
@@ -43,7 +55,8 @@ For example:
 .. code-block:: python
 
    from atlasapiclient import client as atlasapi
-   client = atlaspaiclient.APIClient()
+   
+   client = atlasapi.APIClient()
    client.refresh_token()
 
 This will work for any of the clients included in `atlasapiclient.client`. If 
@@ -53,7 +66,7 @@ Quick Recipes
 =================
 
 Cone Search
-~~~~~~~~~~~~
+-------------
 
 The cone search requires **four parameters**:
 
@@ -76,10 +89,10 @@ The cone search requires **four parameters**:
 
 
 Data for a Single Object
-~~~~~~~~~~~~~~~~~~~
+--------------------------
 
-Get the data
---------------
+Get the data for a specific ATLAS ID
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 .. code-block:: python
 
    from atlasapiclient import client as atlasapi
@@ -88,13 +101,17 @@ Get the data
    client = atlasapi.RequestSingleSourceData(atlas_id=atlas_id, get_response=True)
 
 
-Note: Here we don't parse the config file because we assume you have named yours  `api_config_MINE.yaml`.
-The path to that file is encoded in the `API_CONFIG_FILE` variable in the `atlasapiclient/utils.py` file and parsed by default to the classes.
+.. tip::
+   **What is `get_response`?** This argument just tells the function whether you want to make the request
+   right away as soon as the `RequestSingleSourceData` object has been instanciated. Not doing this 
+   allows you to set up the object, check the payload that was created on instanciation, and then
+   you can get the response by calling the `get_response()` method. If you don't need to check the 
+   payload then it's just easier to do it all in one swoop, but when debugging you don't necessarily want
+   to make API calls so we made it an **opt-in** situation.  
 
[…]
