+++ _sources/index.rst
@@ -13,27 +13,18 @@ This is the documentation for the ATLAS API Client package. This provides a
 convenient way to interact with the ATLAS Transient Server REST-API via Python.
 
 .. toctree::
-   :maxdepth: 3
-   :caption: About
+   :maxdepth: 2
+   :caption: For Users
 
    about
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Docs
-
    users
+   API Reference <api-reference>
+
 
 .. toctree::
    :maxdepth: 2
    :caption: For Devs
 
    dev
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Manual
-
-   API Reference <api-reference>
 
 
