+++ _sources/about.rst
@@ -13,7 +13,7 @@ ATLAS is composed of four telescopes, two in Hawaii and one in South Africa and 
 
 
 Who/what is this code for?
-=====================
+============================
 
 This package allows users of the ATLAS Transient web server to
 query the server without being exposed to code breaking changes to the underlying API.
@@ -28,7 +28,7 @@ It also allows users the make routine queries in a friendlier way, such as:
     If all you want is forced photometry given an RA and Dec check out the `Forced Photometry Server`_
 
 How to get access?
-=================
+======================
 The ATLAS Transient web server is our **production server** for the survey, and
 to manage the load on our machines we cannot give unlimited access.
 
@@ -46,7 +46,7 @@ a **one-time use password** which you will have to change on your first login.
 
 
 Help I forgot my password!
-~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 The server doesn't store your email address so it can't send you an email to reset the password.
 If you forget it for good, you will need to email hfstevance@gmail.com or ken.smith@physics.ox.ac.uk.
 We will then use our admin super powers to reset your password to a **one-time use** password
@@ -54,9 +54,9 @@ which you will then be able to reset.
 
 
 Just Admin Things
-==============
+====================
 Privacy Policy
-~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~
 As ATLAS users your first name and last name will be stored on our server.
 Your email address will be stored on a Google Drive spreadsheet alongside the information
 provided with your Data Access Request Form.
@@ -78,7 +78,7 @@ can access.
 
 
 How to contribute to this code?
-~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 If you would like to contribute to this project you are welcome to
 fork and open a pull request however **we strongly recommend asking
 the dev team (via hfstevance@gmail.com) before you start putting
@@ -92,7 +92,7 @@ If there are good first issues we will flag them clearly.
 
 
 Citing this code
-==============
+====================
 Software in science needs recognition so we can keep funding and human
 resources allocated.
 Please cite this DOI in your TNS reports and papers if you used our code, and thank you!
