### 2026-02-09 #9462 @HeloiseS
Alright! One week late but i've finally had some time.

I made all the changes recommended, they were all great points so no push back on my end.

I made a couple PRs but also a few pushes directly to main!

Thanks a lot @ygrange and  @VasanthRajendran and let me know if there is anything else I can do
