### 2026-02-09 #9462 @ygrange
I can keep it short: LGTM!
