### 2026-02-13 #9462 @Romain-Thomas-Shef
@HeloiseS

A few things to do before I recommend the paper for acceptance:

- A minor extra comment on the paper. Two references in the statement of need are bracketed while there is already a bracket.
  > (e.g. (N. Erasmus et al 2025), (Wet, 2025))
  &
  > (e.g. (Radhakrishnan Santhakumari et al. 2024))
  Please refer to the documentation on [pandoc](https://pandoc.org/MANUAL.html#extension-citations) to correct it.

- Can you create a new release with a new version number? The last release dates back a year and we should create a new version number that includes all the work done during the review.
- Once done, can you upload the last version to the paper to an archive (e.g. zenodo)? Make sure that title, author list, version tag, and the license are correct in the archived release

When this is done I can recommend the paper for publication!
