### 2026-02-23 #9462 @HeloiseS
Alright I think it's done?

The repo is tracked on zenodo so the new release created a new archive of the whole thing including the paper  which you can see here https://zenodo.org/records/18745226

Once you are satisfied I will also update the arxiv version to state that it is published :)

Thanks a lot!
