### 2026-02-26 #9462 @ygrange
Since this all seems to be mostly about publication details and I have the page still open in my browser:

Congratulations on the publication! Reviewing for this journal always makes me appreciate the work the authors put in making the scientific software landscape nicer every day :)!
