### 2026-03-02 #9462 @Romain-Thomas-Shef
Dear @HeloiseS,

I believe now we are all set! I am happy to recommend the paper for acceptance.  Thank you for your patience and for submitting to JOSS! And thank you to @ygrange & @VasanthRajendran for the reviews!
