### 2026-03-04 #9462 @warrickball
Hi! I've suggested some editorial tweaks in [a PR in the software repository](https://github.com/HeloiseS/atlasapiclient/pull/41), which you can review and merge or apply yourself. Along the way, I noticed the following:

* Jack Leland has an ORCID in the paper but not the Zenodo archive.
* You cite the [Zenodo archive for `atlas/vras`](https://doi.org/10.5281/zenodo.14983116) but I notice the authors are "Heloise & Smith, Ken W.". I've suggested adding your surname to the bibliography of the paper (mostly so the pipeline doesn't think "Heloise" and "Stevance, H. F." are different people) but you may want to change it in the Zenodo metadata too.
* As I noted in the PR, Ken Smith is marked as the corresponding author but I suspect it should rather be you. You might want to check that the "equal contribution" field is being used as you intended.

Once all this is resolved I'll launch another test run of the publication pipeline and if there's nothing left we'll be ready to publish.
