### 2026-03-05 #9462 @HeloiseS
I have merged your PR - like I said in the PR ken is corresonding author because i do not have an open ended postion and given the state of the job market it is very likely I will move on from academia - at which point Ken will become the maintainer.

let me know if you need anything else :)
