### 2026-03-06 #9462 @warrickball
Thanks once again to @VasanthRajendran and @ygrange for reviewing and @Romain-Thomas-Shef for editing this submission. JOSS simply wouldn't be possible without its community of volunteers.
